+++ tests/service/ProjectOptionsTests.fs
@@ -1,7 +1,6 @@
 ﻿#if INTERACTIVE
 #r "../../bin/v4.5/FSharp.Compiler.Service.dll"
 #r "../../bin/v4.5/FSharp.Compiler.Service.ProjectCracker.dll"
-#r "../../bin/v4.5/FSharp.Compiler.Service.ProjectCracker.Tool.exe"
 #r "../../packages/NUnit/lib/nunit.framework.dll"
 #load "FsUnit.fs"
 #load "Common.fs"